requirements.txt
# FastAPI and server
fastapi>=0.128.7
uvicorn[standard]>=0.40.0

# LangChain ecosystem
langchain>=1.2.10
langchain-openai>=1.1.10
langchain-chroma>=0.2.4
langchain-community>=0.3.22
langchain-core>=1.2.13
langchain-text-splitters>=1.1.1

# Vector store
chromadb>=1.0.9

# PDF processing
pdfplumber>=0.11.0
pypdfium2>=4.30.0
Pillow>=10.0.0

# BM25 keyword retrieval
rank-bm25>=0.2.2

# Configuration and utilities
python-dotenv>=1.0.0
pydantic-settings>=2.7.0

# HTTP client for async operations
httpx>=0.28.0

# OpenAI SDK (for vision API)
openai>=2.20.0
